Updates to git alias post
I tend to never post if I wait until I feel a post is 'perfect', so if you see a post it's probably something I've written in an afternoon and pushed out when I felt it was 'good enough' not to be embarassed by. This is some edits to the rough post to get us a little closer to something more coherant and easier to understand (hopefully).  Tech debt payment (but for writing)!